+++ tests/koji/test_koji.py
@@ -1,24 +1,8 @@
 import os
 
-from pytest import raises, fixture
-from mock import patch
+from pytest import raises
 
 from pushsource import Source, RpmPushItem
-
-from .fake_koji import FakeKojiController
-
-
-@fixture
-def fake_koji():
-    controller = FakeKojiController()
-    with patch("koji.ClientSession") as mock_session:
-        mock_session.side_effect = controller.session
-        yield controller
-
-
-@fixture
-def koji_dir(tmpdir):
-    yield str(tmpdir.mkdir("koji"))
 
 
 def test_koji_needs_url():