+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- n/a
+
+## [2.7.0] - 2021-06-10
+
 ### Fixed
 
 - Fix usage of `errata` source when the an Errata Tool URL includes a path
@@ -112,7 +116,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial stable release of project
 
-[Unreleased]: https://github.com/release-engineering/pushsource/compare/v2.6.0...HEAD
+[Unreleased]: https://github.com/release-engineering/pushsource/compare/v2.7.0...HEAD
+[2.7.0]: https://github.com/release-engineering/pushsource/compare/v2.6.0...v2.7.0
 [2.6.0]: https://github.com/release-engineering/pushsource/compare/v2.5.0...v2.6.0
 [2.5.0]: https://github.com/release-engineering/pushsource/compare/v2.4.0...v2.5.0
 [2.4.0]: https://github.com/release-engineering/pushsource/compare/v2.3.0...v2.4.0